move scrapejob code
